Consolidated the bulk client into the ForceClient
