+++ src/ForceToolkitForNET.UnitTests/ForceClientTests.cs
@@ -15,7 +15,7 @@ namespace Salesforce.Force.UnitTests
         public async void Requests_CheckHttpRequestMessage_UserAgent()
         {
             var httpClient = new HttpClient(new ServiceClientRouteHandler(r => Assert.AreEqual(r.Headers.UserAgent.ToString(), UserAgent + "/v32")));
-            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v32", httpClient);
+            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v32", httpClient, new HttpClient());
 
            try
            {
@@ -24,6 +24,7 @@ namespace Salesforce.Force.UnitTests
            }
            catch
            {
+               // do nothing
            }
         }
 
@@ -33,7 +34,7 @@ namespace Salesforce.Force.UnitTests
         {
             var expectedResponse = new HttpResponseMessage(HttpStatusCode.OK) {Content = new JsonContent(new {})};
             var httpClient = new HttpClient(new FakeHttpRequestHandler(expectedResponse));
-            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v29", httpClient);
+            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v29", httpClient, null);
 
             await forceClient.BasicInformationAsync<object>("");
 
@@ -48,7 +49,7 @@ namespace Salesforce.Force.UnitTests
                 Content = JsonContent.FromFile("KnownGoodContent/UserObjectDescribeMetadata.json")
             };
             var httpClient = new HttpClient(new FakeHttpRequestHandler(expectedResponse));
-            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v29", httpClient);
+            var forceClient = new ForceClient("http://localhost:1899", "accessToken", "v29", httpClient, new HttpClient());
 
             var result = await forceClient.BasicInformationAsync<ObjectDescribeMetadata>("ValidObjectName");
