+++ src/ForceToolkitForNET/IForceClient.cs
@@ -1,11 +1,15 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Salesforce.Common.Models.Json;
+using Salesforce.Common.Models.Xml;
 
 namespace Salesforce.Force
 {
     public interface IForceClient
     {
+
+        // STANDARD
         Task<QueryResult<T>> QueryAsync<T>(string query);
         Task<QueryResult<T>> QueryContinuationAsync<T>(string nextRecordsUrl);
         Task<QueryResult<T>> QueryAllAsync<T>(string query);
@@ -24,6 +28,22 @@ namespace Salesforce.Force
         Task<T> DescribeLayoutAsync<T>(string objectName, string recordTypeId);
         Task<T> RecentAsync<T>(int limit = 200);
         Task<T> UserInfo<T>(string url);
+
+        // BULK
+        Task<List<BatchInfoResult>> RunJobAsync<T>(string objectName, BulkConstants.OperationType operationType, IEnumerable<ISObjectList<T>> recordsLists);
+        Task<List<BatchResultList>> RunJobAndPollAsync<T>(string objectName, BulkConstants.OperationType operationType, IEnumerable<ISObjectList<T>> recordsLists);
+        Task<JobInfoResult> CreateJobAsync(string objectName, BulkConstants.OperationType operationType);
+        Task<BatchInfoResult> CreateJobBatchAsync<T>(JobInfoResult jobInfo, ISObjectList<T> recordsObject);
+        Task<BatchInfoResult> CreateJobBatchAsync<T>(string jobId, ISObjectList<T> recordsObject);
+        Task<JobInfoResult> CloseJobAsync(JobInfoResult jobInfo);
+        Task<JobInfoResult> CloseJobAsync(string jobId);
+        Task<JobInfoResult> PollJobAsync(JobInfoResult jobInfo);
+        Task<JobInfoResult> PollJobAsync(string jobId);
+        Task<BatchInfoResult> PollBatchAsync(BatchInfoResult batchInfo);
+        Task<BatchInfoResult> PollBatchAsync(string batchId, string jobId);
+        Task<BatchResultList> GetBatchResultAsync(BatchInfoResult batchInfo);
+        Task<BatchResultList> GetBatchResultAsync(string batchId, string jobId);
+
         void Dispose();
     }
 }
